[all] Split Tag & TagSet into separated files and fix missing includes.
