+++ Sofa/framework/Core/src/sofa/core/objectmodel/Base.h
@@ -26,7 +26,7 @@
 #include <sofa/core/objectmodel/Link.h>
 #include <sofa/core/objectmodel/BaseClass.h>
 #include <sofa/core/objectmodel/BaseObjectDescription.h>
-#include <sofa/core/objectmodel/Tag.h>
+#include <sofa/core/objectmodel/TagSet.h>
 #include <list>
 #include <sofa/core/sptr.h>
