+++ Sofa/framework/DefaultType/src/sofa/defaulttype/typeinfo/TypeInfo_Vector.h
@@ -23,6 +23,7 @@
 
 #include <sofa/type/vector.h>
 #include <sofa/defaulttype/typeinfo/models/VectorTypeInfo.h>
+#include <sofa/helper/logging/Messaging.h>
 
 namespace sofa::defaulttype
 {
